MicroService/recommendations/requirements.txt
grpcio-tools ~= 1.30
mysql 
grpcio
